refactor(TripPreviewLayoutBase): Extract common component between print and preview layouts
